Phase05/Test/SearchTest/InvertedIndexTest.cs
using Phase05;
using Phase05.Search;
using System.Collections.Generic;
using Xunit;

namespace Test.SearchTest
{
    public class InvertedIndexTest
    {
        [Fact]
        public void CreateIndexTest()
        {
            var invertedIndex = new InvertedIndex();
            var docs = SampleCreator.CreateStringList();
            invertedIndex.CreateIndex(docs);
            Assert.Equal(invertedIndex.Index, SampleCreator.CreateIndex());
        }

        [Fact]
        public void AddToIndexTest()
        {
            var index = new Dictionary<string, HashSet<int>>();
            index.Add("chocolate", new HashSet<int> { 3 });
            var invertedIndex = new InvertedIndex();
            invertedIndex.AddToIndex("chocolate", 3);
            Assert.Equal(index, invertedIndex.Index);
        }

        [Fact]
        public void AddToIndexTest2()
        {
            var index = new Dictionary<string, HashSet<int>>();
            index.Add("chocolate", new HashSet<int> { 3 });
            index.Add("place", new HashSet<int> { 3 });
            var invertedIndex = new InvertedIndex();
            invertedIndex.AddToIndex("chocolate", 3);
            invertedIndex.AddToIndex("place", 3);
            Assert.Equal(index, invertedIndex.Index);
        }
    }
}
